requirements.txt
slackclient
python-RIBBIT
